chore: remove reporting of attachments to sentry [IDE-1046] (#675)
* chore: remove reporting of attachments to sentry

IntelliJ marked the API used as internal with 2025.1. This means we cannot report it anymore.

In the future, we would use analytics service anyways for error reporting, so they turn up both in datadog and snowflake

* fix: Add INTERNAL_API_USAGES to plugin verifier failure levels

The plugin verifier task was missing `INTERNAL_API_USAGES` in its failure level configuration. This change adds it to ensure that internal API usages are also considered as failures during verification.

This should lead to us catching compatibility issues with internal api changes in newer IDE versions.

Relates to: #None