fix: TextEdits applying to the wrong lines (#697)
The previous code was based on bad logic for creating WorkspaceEdits in the LS.
This LS bug has been fixed, but it isn't compatible with our code which expected the incorrect WorkspaceEdits.
This has been fixed by applying the edits from the end of the file to the start of the file.