fix: remove stale JS which was displayed in HTML
Due to LS not putting `${ideScript}` in `<script>` tags in the IaC HTML, the JS was being displayed as text.
But `${ideScript}` is in `<script>` tags in the Code HTML, leading to an issue.
Luckily we can skip the issue, as the JS code injected is all stale now anyway.